requirements.txt
cython
thriftpy2
ptsd
pytest
